feat: implement mouse click overlay processing for video and GIF captures
